chore: restructure core/hub/
