+++ wallets/core/src/builders/mod.ts
@@ -1,4 +1,4 @@
-export type { ProxiedNamespace, NamespaceInterface } from './types.js';
+export type { ProxiedNamespace, FindProxiedNamespace } from './types.js';
 
 export { NamespaceBuilder } from './namespace.js';
 export { ProviderBuilder } from './provider.js';
