+++ wallets/core/src/hub/namespaces/mod.ts
@@ -1,0 +1,10 @@
+export type {
+  Subscriber,
+  State,
+  ActionsMap,
+  AndUseActions,
+  Context,
+  Actions,
+} from './types.js';
+
+export { Namespace } from './namespace.js';
