+++ wallets/core/src/hub/provider/mod.ts
@@ -1,0 +1,9 @@
+export type {
+  ExtendableInternalActions,
+  CommonNamespaces,
+  State,
+  Context,
+  ProviderBuilderOptions,
+} from './types.js';
+
+export { Provider } from './provider.js';
