+++ wallets/core/src/hub/store/mod.ts
@@ -1,0 +1,8 @@
+export {
+  guessProviderStateSelector,
+  namespaceStateSelector,
+} from './selectors.js';
+export type { Store, State } from './store.js';
+export type { ProviderInfo, ProviderConfig } from './providers.js';
+export type { NamespaceConfig, NamespaceData } from './namespaces.js';
+export { createStore } from './store.js';
