+++ wallets/core/tsconfig.build.json
@@ -1,12 +1,10 @@
 {
   // see https://www.typescriptlang.org/tsconfig to better understand tsconfigs
   "extends": "../../tsconfig.lib.json",
-  "include": ["src", "types"],
+  "include": ["src"],
   "compilerOptions": {
     "outDir": "dist",
-    "rootDir": "./src",
     "lib": ["dom", "esnext"],
-    "jsx": "react",
     "target": "esnext",
     "module": "NodeNext",
     "moduleResolution": "NodeNext",
